authorization user check on update and delete

+++ module-25-advanced-rest-api/api/controllers/feed.js
@@ -109,6 +109,11 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized!');
+            error.statusCode = 403;
+            throw error;
+        }
         if (imageUrl !== post.imageUrl) {
             clearImage(post.imageUrl);
         }
@@ -134,7 +139,11 @@ exports.deletePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
-        // check logged in user
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized!');
+            error.statusCode = 403;
+            throw error;
+        }
         clearImage(post.imageUrl);
         return Post.findByIdAndRemove(postId);
     }).then((result) => {
